sreality_scraper/requirements.txt
scrapy==2.9.0
psycopg2-binary==2.9.6